Uploaded Data Files
FullDataset_Final is the file with all the CT-level variables that need to be imported to R for analysis. 
FullDatasetNew is the one generated by R that contains both the original variables and the OLS residuals. It needs to be imported to ArcGIS to see the spatial autocorrelation of the OLS residuals.
SpatialDataset is the shapefile generated by importing FullDatasetNew to ArcGIS to test spatial autocorrelation of OLS residuals. It is then exported from ArcGIS and imported to R to calculate Spatial Lag Regression. 
FullDatasetSpatial is the one generated by R that contains both the original variables, the OLS residuals, and the SLR residuals, which is imported into ArcGIS to see the spatial autocorrelation of the SLR residuals.